introduced a namespace "Output" and methods print
This is now the preferred way to write to console and files. Read its
documentation! The old macro OutPut is still supported for backward
compatibility but it is deprecated, do not use it any longer. Also don't use
cout or printf directly.

Use Output::print<x>(...) where 'x' is 1,2,3,4, or 5. The larger the number
the larger the verbosity level has to be in order for this to be printed.