first try to use umfpack with 'long' instead of 'int' as structure indices
The structure class only supports int for its row and column pointers. For
larger matrices we need to call umfpack with long rather than int. The idea
is to copy these arrays locally to std::vector<long>. This is not a nice
solution. We need to define structures with long, int and possibly also size_t
as index type (template?).