Bugfix in Assemble3D (with LocalAssembling object).
The loop which took care of setting Dirichlet rows in the global matrix assume that there was only one entry per row. This is not necessarily the case, when the full structure gets assembled. Assuming just one entry led to no ones on the diagonal being set. Fixed it by searching through the row for the diagonal element.