included Pardiso
this uses an old version of pardiso which is not strict about the license. You
have to enable openmp to use it, otherwise you will get an error trying to use
it. So set PARMOON_PARALLEL_TYPE in your CMakeCache to e.g. OMPONLY.

There is a newer version of pardiso also in the repository (in the mumps
directory), but that requires a license file which we can not put in the
repository.

Note that older versions of clang don't support openmp, which results in an
error while running cmake. You have to switch to a newer version or to a
different compiler.