Promising tests on TNSE3D.
4 example files coming from MoonMD were added and tested. Example_NSE3D has been adapted in consequence.
For default solving parameters (nstype 1, vspace 12, timedisc 2,...), errors between computed and exact solution are ok. Order of magnitude 1e-15 for linear functions, 1e-6 for polynomial and exponentials, 1 for 4th degree polynom in time. Only problem noticed is with example 102 AnsatzLinConst, which is linear and should have a very small error, the error is 0.5. Deeper work needed on this point.