VERY IMPORTANT CORRECTION OF MPI CODE IN TNSE3D.
Correction to be taken into account in all non-linear problems (time-dependent and -independent problems) in MPI case : remember to CommUpdate the solution vector whenever it looses its consistency (eg after copy_nonactive). This ensures that the defect and other terms involving a multiplication with the solution vector are correctly computed!

TNSE3D running with MPI gives now similar results as with SEQUENTIAL. Minor and until now not critical differences still appear but it is not clear yet if there are due to the solver differences or to something else...

It is possible that the code needs additional CommUpdate at some points (eg. old_rhs and rhs after copy_nonactive in assemble_rhs()). This is still an open question.

Anyway, current examples (101 to 106) don't require futher adjustments...