update of the ParameterDatabase
When reading the documentation of a parameter into a database you have to
precede it by '#'. If you would rather keep the default documentation of that
parameter (if it is defined in some default_database method), you simply write
'##' in front of every line of the documentation. E.g. in your input file

## some description which will be ignored by ParMooN
outfile: my_outfile.out

Reading this file will leave the parameter outfile with its default description