forward declarations instead of includes in Solver.h
I had to change unique_ptr to shared_ptr for the member direct_solver. It
should be unique_ptr. However to use a unique_ptr the compiler needs to know
the complete type for the destructor. The default destructor which the
compiler provides is inline. So this will not work. You can make the implement
the destructor in the source file (e.g. Solver<L,V>::~Solver() = default;) or
use a shared_ptr (which I used now).