changed the output of iterative solvers
this is no longer done in the Solver class, but in the class IterativeMethod.
There, during every call to IterativeMethod::converged(..) some information can
be printed. Also we now store the previous residual to check if the convergence
is slow. Here we could have another factor as a parameter in the database.

Also there is more output, if there is any output. Namely the reduction
compared to the previous and the initial step.