little refactor of Multigrid class
it now stores a parmaeter database and is constructed in essentially two steps.
The constructor is only setting some variables. Then you have to call
initialize. The reason I did this is, that this way one can ask the multigrid
object for the number of levels and is_using_mdml which is important building
the matrices in the classes where multigrid is used (the system classes).