updated the output of IterativeMethod::converged, some documentation
also I had to relax the darcy test slightly. This affects RT2 on triangles,
using the lsc preconditioner. It takes way too many iterations. But this has
not changed, compared to earlier version. But now the solver stops as soon as
the residuals are increasing. The differences are very small (1.0e-9).