updated cd2d test to inlcude more elements and solvers
P1,...,P6, Q1,...,Q8. Note that the higher order ones do not work weel with
iterative solvers. Here we test with a direct solver and with fgmres. As
preconditioners we test ssor and multigrid.

The old test using algebraic flux correction is still there and running.


Also updated the darcy2d test. There a 'return' statement was removed
effectively enabling a number of tests.