updated including of external libraries
we no longer inlcude the external libraries by hand. Instead we use the
variable PETSC_LIBRARIES which lists all the libraries we need.

Also instead we check in the PETSC_LIBRARIES if the required libraries are
present. This is a bit of a hack. If the libraries don't have their common
name, this might not work.