Improving the method norm() of BlockVector to unify SEQ and MPI cases.
The method norm() is now called in all solvers, replacing norm_global().If the communicators are provided, then norm() understands that it should call norm_global(), if not, it will call the standard sequential norm with a warning to tell the result is certainly not correct. It is maybe temporary and needs improvement.