Added the flow around cylinder problem in the folder data/.
3 files : PRM, GEO and .dat

As the flow around cylinder is an important benchmark problem which is necessary to check our code, it is important to have it in the main ParMooN folder. So far, the files correspond to steady-state, 3D with hexahedra benchmark problem. Later, we can add other geometries (2D for instance) and add/adapt input file to use it with time-dependent nse3d.