FGMRES is now parallelized.
Just put the comm updates, the norm_global and the dot_product as in bi-cgstab and richardson. No change in the methods about Givens Rotation or solving the Hessenberg matrix. It seems that the comm_updates, the global dot and norms are enough to give the same Hessenberg matrix for all the processes...

The comparison with SEQ gives correct results with one particular case. Further tests are necessary to validate the solver.