Finished validation of Test Poiseuille.
Actually, the dimensional NSE was hand tested by forcing mu and rho coefficients in the assembling routine.
For different reynolds numbers, dimensional and non dimensional nse give similar results but not exactly the same.
The results are very close to 0 and there is no iterations so it is difficult to conclude if the dimensional nse works well.

Next step consists in adding a new example where we will better compare dimensional and non dimensional nse.