Now fixed.
FROM NOW ON, in the input file, the user should always give consistent rho, mu and reynolds number, even if it is redundant: in this way, the parameter "dimensional_nse" can be changed without having to adapt the code in the example files. Workarounds can be later found...

Now, all tests for NSE2D are validated with the new layout and with both dimensional and nondimensional NSE, in the program multiphase.

Next step: validate the parameter coupling_nse_cd, by using the standard test examples of CD2D and generating the examples for NSE2d to obtain the velocities corresponding to the examples of CD2D.