Added a new test example in CD2D. And first validation is ok.
This is to test the combination POISEUILLE+2FLUIDS, (nse_example=11 and cd_example=11...pure coincidence...no, really!).

In this commit, a first test runs problemlos: the concentration is 1 everyone, there is only one fluid submitted to the classic poiseuille flow, all the couplings are activated and dimensional_nse=1, but all property fields are 1.
Converges in one iteration and reproduces the expected velocity field.

Now that we have this, next step consists in changing this cd2d example file, by introducing boundary conditions for the 2 fluids, as a stratified flow submitted to the same poiseuille field, and with different viscosity and density and see what we get...