BEGINNING OF NEW WORK PACKAGE: time-dependent multiphase program!
It was found that the steady-state multiphase model was difficult to check and is not really relevant.
The time-dependent multiphase model has a lot of benchmark problems which makes it easy to check and validate.

Started a new main_program, added in the CMakeLists, added a new input file, which is so far exactly the same as multiphase2d.dat (steady-state).

Next steps: write the program to solve TimeNSE2D with the current layout of multiphase and test with examples.

Remark1= no input file for TNSE2D was found...but it is not the focus of the branch so I wont do it now.
Remark2= the steady state multiphase is left as it is in the last commit: one-way coupling is validated but not the 2-way coupling. No examples were found to check the code, so I will leave it like this and focus on the time-dependent where I can check my code with benchmark problems.