CODE VALIDATED.
The TNSE2D algorithm in the TimeMultiphase2d gives same result as the standard program tnse2d for both tests 0 and 1.

+ Added the time parameters in the input files.
What was done so far is just a copy paste of the tnse2d algorithm...so nothing exceptional happens.

Next step: Adapt assemble methods to solve the DIMENSIONAL NSE!!