Assemble initial time with fluid properties is now implemented!
A compiling error (multiple definitions) was appearing when I included header file TNSE2D_FixPo.h in the Time_NSE2D.C to be able to include the routine corresponding to dimensional NSE.

It was strange because I did the same with NSE2D and had no compiling error (namely, I was able to make it run by simply adding #include <NSE2D_FixPo.h> in the NSE2D.C file).

This difference of behavior has now an explanation: the NSE2D_FixPo.h has only declarations of routines, whereas TNSE2D_FixPo.h has declaration of routines AND DECLARATION+DEFINITION OF SOME VARIABLES...these variables are not called anywhere in the program, so they can be considered as useless, but still they can be helpful because they indicate how the local assembling properties must be set to fit with corresponding routines.

So when we #include TNSE2D_FixPo more than once, we have multiple definition of these variables...which is not the case in NSE2D_FixPo because there are only routine definitions...

In order to workaround this situation (not being able to #include TNSE2D_FixPo.h and not wanting to suppress all defined variables), I added methods in the LocalAssembling object which reads a string (so to write strings in Time_NSE2D.C) and updates the local assembling to register the corresponding routine.

It is not very nice..but does the job conveniently with very few effort and is understandable..

Next steps: make the code cleaner and test the assemble_initial_time_withfields, by comparing with non dimensional NSE.