Working commit: added TCD2D in main program + definition of rho and mu FEFunction + reading "dimensional_nse=true" in main program and TNSE2D.C + improved output messages of the 2 localassemble methods...
Remark: the routine assemble_initialtime_withfields was not changed yet here.

Next step: validate the whole worked code by comparing dimensional_nse = true and false. This will validate the implementation of the call hierarchy of the new routine and methods in TNSE2D and LocalAssembling2D as well as the detection of fluid properties.