We have to change the assemble routine for RHS in order to take into account rho.
Unlike the steady-state NSE, where RHS is assembled together with A matrix and thus can read rho, in TNSE2D, the assemble routine for RHS is separated and dont have access to parameters[2], which is rho, so no way to have it multiplied by rho in the example class.

A new assemble routine called TimeNSERHS_dimensional was created, as well as a new method assemble_rhs_dimensional. All necessary things were added in the header and source files (Local Assembling.c, TNSE2D_FixPo.h and .c , Time_NSE2D.h and .c).

Next step: add it in the main program, and test.
Then, work on the routine itself to get the rhs multiplied with rho correctly.