CODE VALIDATED!!!!
FROM THIS COMMIT, THE DIMENSIONAL TNSE2D IS OPERATIONAL AND GIVES SAME RESULT AS IN STANDARD NON DIMENSIONAL TNSE2D PROGRAM FOR THE EXAMPLE BSP1 (sincos) with different end times and reynolds number = 50 (in the dimensional we use rho=100 and mu=2).

Next step: Make the example code cleaner and maybe implement the other example linspacetime...or not?