STARTED TO ADD TCD2D IN THE MAIN PROGRAM.
I suppose that the tnse2d is ok and that the validation of the code I did earlier is sufficient.

Now focusing on tcd2d, first step, just solve any example and compare with standard tcd2d results to validate its implementation in my multiphase main program.

if that is ok in the next commit, then I will start a new assemble method which takes the velocity from TNSE2D as an input convective velocity in the TCD2D object...