CODE APPROXIMATELY VALIDATED!!!
THE TCD2D GIVES SAME RESULT AS TCD2D MAIN PROGRAM FOR EXAMPLES 0, 1, 2....but there is only one non linear iteration, which is weird because when coupling is deactivated, then TNSE2D makes 2 non linear iterations..

Further tests are needed to be sure that the code is validated. It is namely important to validate the tests where several non linear loops are done, which means that tcd2d  is several times  assembled, solved and descaled in the same time step...this is next step.