Commented out all the variables of these 3 header files.
Why? Because in header files, there should not be DEFINED variables, only DECLARED. These variables were declared AND DEFINED. Thus, when #including <these headerfiles.h> anywhere in the program, we make a compiler error "Multiple definitions". Now, we dont have this problem anymore.

Why not suppress these variables instead of outcommenting them? These variables are not called anywhere in the program so indeed, it is possible to just delete them. Make check will still work fine. But they describe how the local_assembling object must be set in order to work with the corresponding assemble_routine, so it can be helpful to keep them and not throw them away...although we can admit most of the assemble routines are not tested at all in the code, so we dont even know if they work....