STARTING WORK ON TCD2D.
TCD2D assemble method has to be split in several parts: assemble_rhs, scale_matrices, assemble_stiffness and descale_matrices.

These parts have to be correctly integrated in the loops: assemble_rhs once per time loop, before the non linear loop, scale_matrices (and reset matrices) and assemble_stiffness inside the nonlinear loop, which can be solved several times if needed, and descale_stiffness at the end of the time loop, after the non linear loop.

To start, it is necessary to validate the code re-working. For that, the result of a test example (for example 2) is compared between multiphase and tcd2d. Max non linear iteration is set to 1, so that the code should always give the same result when tcd2d is solved just once in the non linear loop.

Then, step after step, if everything works fine, we just need to put max non linear iteration to more than 1, to see if the code works correctly.

Next step: Split assemble method of tcd2d in several parts.