SEGFAULT with the new assemble_stiffness_matrix_alone.
Needs to be corrected. To check the code while debugging, just comment in or out the line "tc2d.assemble_stiffness_matrix_alone();" int he main program.

When it is out, the stiffness is assembled just once in the rhs routine and we get the correct results (number of nonlinear iteration=1). If it is in, the stiffness should be assembled alone, and here is the segfaut.

Next step: Debug the assembling of the stiffness matrix alone. Namely, rhs_entries should not be used at all, only A should be reassembled.