CODE VALIDATED!!!!!!!!!!!!
Added a new assemble routine where only A is assembled, it is called LocalMatrixA_alone, and is defined in the ConvDiff2D.h and .C

In the assemble_stiffness_matrix from Time_CD2D, I used setAssembleParam to set this routine in the local assembling object.

By express-testing in main program, the algoritm gives the correct result, as well as with several iterations in the non linear loop!!!

This means that in each non linear loop, the convection diffusion seems to be correctly set and solved, namely, the stiffness matrix is reset, reassembled and rescaled, while rhs and mass matrix stay the same.

Next steps: Intensive testing with several examples, several non linear loops and if everything is ok, clean the code! and commit