CODE VALIDATED!
The current code was compared with standard TCD2D for examples -1,0,1,2 and for 5 time steps and with repeated nonlinear iterations.

At each non linear step, the convection equation gives same results and each time step, same solution as in TCD2d.

It is concluded from this small test campaign, that the current code is validated and that tcd2d in multiphase program is solved correctly.

Next step: see next commit