Tests OK.
There is however this small problem with the residuals: the calculated L2 and H1 residuals with time are not the same between standard tcd2d and multiphase :/

This has to be fixed...but later..

Next step: just check Paraview to see if solution of convection equation between standard and multiphase programs are the same for this example ExpSinCos.

After this, make the same with the other examples.