EXCELLENT!!!
All the test examples of TCD2D (linear space time, expsincos, sinsinsin, sincos) were addded to the multiphase project in order to be changed in the next commits to be convected with velocity fields from navier stokes.

So far, they were just tested with coupling_nse_cd=true and the standard assemble method of Tcd2d (assemble_stiffness_alone) and compared to tcd2d standard program, on 15 time steps.

All the results are the same, so the code is validated until now. Paraview images were also checked-> successful.

IMPORTANT NOTES: A DIFFERENCE HAS BEEN OBSERVED, in tcd2d program, there is always one more VTK than in my program, this is normal, it comes from the call of tcd2d.output() before solving the problem, so it is just an initial image with c=initial (=0 in most of the cases). So dont be surprised if you see a slight difference between the VTKs. Another thing is this L2 and H1 norms error with time, they are almost always the same as in standard program but in one case I remember it is somehow different...I wrote it in a previous commit...needs to be checked in the future...

Next step: just activate the assemble_stiffness_alone_withfields method in the main program to test it as it is now...there should be NO difference, because the convection field is not used at all so far in the example classes. Just to be sure.