Code not validated.
Everything in this commit is ready to test. Just change the example tcd2d (1_TestExampleTCD2D.h) in this way: exchange b1 and b2 values from (1,-1) to parameters[][]. If it is (1,-1) you get the same results as the standard tcd2d program. If it is parameters, not the same results, so code not validated.

It seems it comes from the parameters values, which are not calculated correctly...always this problem.

Next step: investigate it and solve it. Then the code must be validated. Then, the same will be repeated with the other examples (just put parameters[][] in all the examples instead of hard coded b1 and b2 coefficients).