PROBLEM SEEMS TO BE SOLVED!
The problem was that two-fold: the right hand side depended on the convection field, so it has to be assembled with the convection_field and the initial assembling of rhs and stiffness matrix also had to be assembled with the input convection_field.

Now 2 methods do this (just a copy paste + few lines to get the convection field parametrised). Quick test on one example with 1, 3 and 5 time steps show agreement with the standard TCD2d program.

Next step: Intensive testing with other time steps, and other examples to be sure everything works fine + code cleaning!