CODE VALIDATED!!! Coupling NSE > CD WORKS FINE.
Tested two examples with 5 time steps and coupling_nse_cd=true. See the files.

You can use this commit to test again and verify validation. In the main_program, put (tcd example=11 and tnse example =12) or (tcd example 12 and tnse example 13) and run the simulations (coupling_nse_cd should be true).

This should do all the assemblings with the convection field from Navier-Stokes which has been specifically calculated to match the correct convection_field. Then run simulations with the standard tcd2d main program with the corresponding examples (0 and 1) and you should get the same results.

This is like a milestone commit.

The code is not so clean so far...some day I should make it nicer.

There are still 2 more examples that have not been tested and which should be tested, namely SinCos and LinearSpaceTime..but I suppose the current tests are relevant enough to validate the code...and I dont want to make extra copy paste and make the code too heavy...when we have a better example and testing structure, it can be done.

Next step: pause...and think about next step...which is certainly to implement the coupling cd > nse.

Last thing: all the tests have been done with dimensional_nse=true but with eveything=1, reynoldsnumber, rho and mu... it has not been tested yet with other values, and this may be an important thing to check...comes also with the preparation of the next steps.