I just played around with the tests...repeated tnse2d tests with dimensional_nse and compared with standard tnse2d...
Next commits will be playing around with tcd2d and tnse2d examples to get some interesting results and testing cases...

For example: take the velocity field (1,-1), I think it is tnse2d_example=12. Create a tcd2d example where a column of phase 1 is located in the square box, surrounded with phase 0. It means only to define initial concentration field. Think of the boundary conditions...and try to convect it with the constant velocity field (1,-1) and see what we get...the idea is to understand how to define correctly the tcd2d example, initial function + boundary type and value. Correctly means here to emprison 2 phases in the box and submit them to a convection field...see if it works, then if it is ok, we can think of creating a new tnse2d example where gravity is taken into account, velocity is 0, wall boundary conditions = homogenous dirichlet...

Lets start with the new tcd2d example.