Successful test : also compared with tcd2d (velocity field = (1,-1)) and same results viewed in paraview over 50 time steps.
The initial field is ok, and it is correctly convected in direction (1,-1).

But it doesnt look good at all: grid too coarse, spurious oscillations, the need of stabilization is obvious and essential, the concentration diverge and doesnt stay between 0 and 1, probably because of inappropriate boundary conditions...and so on..

Many improvements are needed. But first thing is that it is ok and I guess if we refine mesh and time step then it will be better. But Stabilization will be needed anyway.

Now next step = create a tnse2d test file to define NS equations: initial velocity = 0 everywhere, boundary conditions = homogenous Dirichlet everywhere, reynolds number = 1, right hand side = gravity (and check it is multiplied with rho in the assembling routine...this should be the case).