VALIDATED!
As expected, the coupling nse > cd didnot change anything, because velocity remains 0. Concentration is solved and is now e-15 instead of rigorosly 0, but the results are anyway those expected.


Now that we have this interesting and relevant test case (tnse2d example = 14 and tcd2d example =15), we can work on the coupling cd > nse.

The main work is in the main program, make sure that the concentration field is applied to the property fields rho and mu.

Next step: Implement (=uncomment =D) coupling cd > nse features.