Run a long simulation and saved the output, vtks and images in a folder simulation_runtest.
First thing, tried to make a video from the 1500 images, but problem with video reader...comes from video reader AND/OR from encoder...needs to be fixed.

Second, although refinement was 7 and time step 0.001, loop didnt converge, all was run with max_nonlin_iteration=2....but it convected as expected. Pressure drop because of water column created velocity on the bottom part. and pressure above also created vertical velocity. Difference of densities is factor 2...anyway, problem is not well posed.

Needs several things: first= stabilization, second=check well-posedness of problem and algorithm. And numerical condition for convergence.

Here is an input file to play around with the current version of the code.


Next step to clean up the program and further analyze: implement a 2nd test case, different from the water column. This should be the circle convected with a rotational velocity field, benchmark problem in a thesis. This should enable to further check/validate/correct the code.


Another important point, after having implemented this...is to check if we move out convection equation for the non linear loops...it seems it has no effect, only extra unuseful computational cost + destabilizing more navier stokes....if we take it out, the phase fraction is then purely convected and property fields are updated once per time step....but first, lets define the convected circle case and we will test on it.