Test validated!!!!
The semi-circle is convected as expected.

Only problem is oscillations and then solution explodes. It is normal, space and time grids are coarse and no stabilization.

Here is the input of a test with finer time step and space grid.

Next step:...pause and think....maybe activate the coupling cd>nse and see if convergence is destroyed...and then test the program by moving out the Convection equation from the non linear loop...but anyway later or soone r SUPG has to be implemented and tested for both nse and cd...