COMMIT TO DEFINE NEXT STEP.
(this is just an empty commit, used to write here the situation).

The previous commits and simulations were partially successfull. We faced a major problem : the iterations did not converge!

After reanalyzing the equations and the algorithm, it was found that the mass matrix was assembled just once in the whole code, with the initial rho field. BUT the rho field changes, so it has to be assembled in every non linear iteration!

This is the main point missing. And it is the only point identified which could explain the non convergence. So using the code in its current state is of no use, because it is not correct..even if it gives nice pictures.

Next step is = think of the best way to reassemble the mass matrix (should it be included in the assemble_nonlinear_term_withfields method?) and do it. Then test again.

Once we have this, check if we have convergence...and the step afterwards will be the stabilization (SUPG and Algebraic Stabilization)...