Starting implementation of an assembling routine for the mass matrix with rho field.
It is not operational yet, two things are missing: the configuration of the LocalAssembling object when rho_field != nullptr and the implementation of the corresponding localassembling configuration AND new assembling routine. So far, I wrote TNSE2D just in order to make it compile. It has to be changed.

But first, needs to set up a test case to validate the new assembling routine. Needs to find which example I can use and how to test the code.

Next step: Set up a relevant test to validate that the new assembling routine for the mass matrix works correctly.