Bug fixed.
It was due to the initial phase field. It was not unity anymore, beginning from commit 1902.

It was taken as the initial solution of TCD2D to test the coupling CD>NSE...and this doesnt match the benchmark problem Bsp1 for TNSE2D.

This was fixed by putting a condition to detect if coupling cd > nse is active or not. This way, the tests of tnse2d can be confidently and correctly run.

Now that we corrected this bug, continue to set up a test environement to validate the implementation of the new assembling routine for mass matrix.