Finished to set up the code for tests.
This is a brutal set up, just commented out the Convection part to speed up solving, added line to re-set phase field to unity. These lines HAVE TO BE REMOVED WHEN THE TEST IS OVER.

Now, the dimensional NSE gives the same result as the standard TNSE2D program for example BSP1, also when going through the if conditions, because all couplings are active.

For the test of the mass assembling routine, just have to add the line in the main code and see if we obtain the same result.

Next step: Implement the mass assembling routine and configure the local assembling object appropriately.