Working commit.
Code written so far has been partially checked in the following way: in the assemble_method, the local assembling object is created and configured, and it calls the assembling routine TimeNSType1GalerkinMass_dimensional, which is empty. But, I deactivated the reseting of the mass matrix...and I got the right results.

It means that the assembling configuration, methods and objects are correctly defined, at least they compile and dont throw SEGFAULT.

And it means that the call of the assemble methods works, when the assemble routine for the mass matrix is empty and when mass matrix is not reset.

When I comment the reset of mass matrix, it doesnt converge anymore...because the mass matrix is reset to 0, and it is reassembled...with an empty routine!

Next step: just to enjoy that the code works so far, and to be even more sure, print out some info from the assemble routine to see if the call works fine...