Quick validation is OK!
Written the assembling routine for the mass matrix with rho field. Reset mass matrix before each re-assembling.

Cleaned up the code to get rid of compiler warnings (unused variables, set but not used).

Tested rapidly and obtained same result as standard tnse2d program. Printed out the rho field to check it, it was correct. So it means that the coupling cd>nse is correctly re-assembling the mass matrix with rho.

Next step: think of other relevant tests to be sure that it works fine...and then ultimate test = re-make the 2-phase flows and check convergence of this Volume Of Fluid algorithm....