Some ideas to debug.
Most importantly = find an example where rho and mu are not homogenous and check when there should be convergence! and why my program fails.

Check following things to have more ideas about convergence of VOF model : check equations in the Master thesis Sweden, in the PhD thesis of level-set methods, on french course on variable density Navier-Stokes, and look on the internet on VOF algorithm or VOF convergence...



Anyway, it is essential to find a benchmark problem where the density is not homogenous and which can be reproduced in ParMooN...and look at convergence....maybe it is a good start to see french paper.