The Coupling CD>NSE HAS TO BE ALL DONE AGAIN!!
Here are new elements to study the code and solve the convergence problem.

First: the A matrix was reassembled each time in the non linear loop with the INITIAL rho and mu! Because after the condition "if coupling cd>nse=true", there was a "if dimensional-true" instead of ELSE if!! The word "else" was missing and has to be added...this was the first thing and it has been done in some previous commit...


Then, however, there was still no convergence...so I just commented out the whole block "if coupling cd > nse=true" to see if there is some change, but still the same situation!

So I put "coupling cd>nse=false" and it worked! this line was just used once at the beginning of the code, when initializing the phase-field!

It seems that the problem doesnt come from the non linear loop or from assembling methods, but really from the discontinuous phase field injected in the initial equation of NSE...if NSE is assembled with inhomogenous rho and mu, it is not converging...

Another thing, is that the rho and mu are not updated correctly, because two new fields are created, and not used in the assembling of the right hand side!!

So the whole coupling cd>nse has to be done again..starting from the initial assembling and then the update of the fields... next steps should focus on those 2 points