This is a commit to prepare meeting with Volker.
Here we can see a quick test of the dam break, with one time step 0.001, verbosity =1, all the couplings are true.

When fluid properties are set to one, the 2 fluids have the same, converges directly! when not, iterates with the same residuals.

Just to illustrate the current state of the code.

Next steps to correct this should be, as written in previous commits, to write rigorosly the problem and to reprogram the coupling cd>nse, because the rho and mu fields are not updated correctly...but still, we have to be sure that the problem is correctly written before.