Algebraic Flux Correction is already implemented!
It was found that the FEM-FCT was already  implemented for tcd2d (crank-nicholson). The parameter Algebraic_flux_correction has to be put to 1 in the input file.

Next step: add the call to function "do_algebraic_flux_correction" in the assemble method with convection field in tcd2d class. and test again the dam break to see if stabilization works