FIRST TEST OF CD>NSE COUPLING IS POSITIVE!
The following test combination was done: NSE_Example=17 and CD_Example=18, corresponding to the Poiseuille flow with a variable viscosity (mu = 1000*y+700).

First, a TCD2D-EXAMPLE was generated to recreate this field. Then, in the multiphase program, rho was put to 1 everywhere, and the program was launched in two-fold: first, without coupling (cd>nse=false), in this case, mu was assembled from the value given in the example; second, with coupling cd>nse=true, in which case, mu was assembled as a Param equals to tcd2d.get_solution() given into the assembling routine.

Same result was obtained, meaning that the mu_field given in the assembling routine of TNSE2D was correctly read.

Next step, just try to activate the coupling nse>cd....this should change a bit... and later we will implement the second example with exponantial