Reworking TCD2D in the multiphase program.
2 things are considered: how to solve properly tcd2d (there seems to be incoherence with the assembling routine+ no good management of the coupling booleans) + stabilization!

To manage this, reworked the example "semi-circle". It will be used : 1) first to assess stabilization in the standard tcd2d program, 2) then to reproduce correctly the example in my program without and with coupling, 3) last, implement stabilization in my program.

Next step: Introduce the rotating convection field in the TCD2D "Semi Circle" example. ATTENTION! It means that
when testing the couplings, I have to uncomment/disable the rotating field written inside the example :/