OK FOR ROTATING SEMI-CIRCLE!
The rotating field was included independently from the coupling with NSE (ie. directly in the example file) and tested in the standard tcd2d program:
1) AFC works only with Ansatz_order=1, fast and precise 2) SUPG works with order 1, but less precise, 3) SUPG also works with order 2, but too long and increased errors.

So it looks like we have to achieve order 1 in the multiphase program for the TCD2D equation.

Next step: rewrite tcd2d in my program to solve it correctly with my assembling routines + test some stabilization to try get the same results as in tcd2d, independtly from TNSE2D