Corrected the name of an example. Added a small paragraph to better manage the usual examples used to validate the code (10, 20,21,30,31, and now 32 BELTRAMI!).
Validated the code by appropriate comparison with TCD2D and TNSE2D Main programs, as well as with and without couplings.

The code is a bit better now, and ready to make further progress with TCD2d (rewrite program, be able to execute it independently from TNSE2D, obtain correct results with the combination "my assemble routine+SUPG and AFC" and compare with standard tcd2d).