THIS IS A TEST COMMIT.
In this commit, the details of the coupling nse>cd was tested: in the example file, the values for "parameters[i]" and the real values of the convection fields are printed in the output file.

When convection field AND scalar field are P2: the values match up at all quadrature points. But when scalar field is for example P1 or P3 while convection field is P2, there are some (sometimes big) differences at the evaluation of "parameters[i]" at the quadrature points.

This commit is very important to confirm that WHEN CONVECTION_FIELD AND SCALAR FIELDS HAVE DIFFERENT ORDERS, THE EVALUATION OF THE CONVECTION_FIELDS AT THE QUADRATURE POINTS JUST GIVE RUBBISH!

...hence the necessity of interpolation :/

To return to previous configuration, just up to the precedent commit.

Next step will be probably about thinking of a solution for this interpolation (doesn't work for P1>P2 interpolation + even if works, excessively long).