Continue testing of interpolation and assembling of convection field in TCD2D...
Weird, the interpolation P2-P2 works but all the values going to the assembling routines are e-300....looks like there is a bad management of ptrs.

Needs to be corrected.

Next step: manage at least once a proper interpolation + injection in assemble routine. And compare P3, P2, and P1 interpolation with no interpolation if the interpolation works OK...the long computation time will be handled later.