IMPORTANT COMMIT: REWROTE THE COUPLING AND INTERPOLATION NSE>CD!
The interpolation is now correctly done, there was a misuse of memory in the "if(convection_field)" condition, which made any interpolation fails.

It has been corrected by the introduction of 2 vectors<double> as members of TCD2D to save the values of the convection FE_Function.

The results of this interpolation was examined through the example 22 (rotating semi-circle) where the analytical value of the convection fields x and y were printed out and compared to the values of the param "tnse2d.get_velocity".

Now, the values match and are correct, even when Ansatz_order=1 or 3 (P1, P3), while Velocity_space remains P2.

Next steps: relaunch the bunch of all the examples + correct the coupling TNSE2D in the same way (interpolation+ correct the assemble routines + check on an example the values of analytical viscosity in comparison to viscosity calculated with tcd2d.get_velocity).